move funnctions to utils
